+++ StreamProcess/src/main/java/applications/bolts/transactional/tp/DispatcherBolt.java
@@ -37,9 +37,9 @@ public class DispatcherBolt extends filterBolt implements Checkpointable {
                 System.nanoTime();
             }
             String[] token = raw.split(" ");
-            Short time = Short.parseShort(token[0]);
+            Long time = Long.parseLong(token[0]);
             Integer vid = Integer.parseInt(token[1]);
-            assert (time.shortValue() == Short.parseShort(token[0]));
+            assert (time.longValue() == Short.parseShort(token[0]));
             this.collector.emit_single(POSITION_REPORTS_STREAM_ID,bid,new PositionReport(
                     time,
                     vid,
